단건 제거 API -> DELETE - /todo/{id} 완료건 제거 API -> DELETE - /todo/complete

+++ pilot-todo/pilot/src/main/java/com/example/pilot/todo/presentation/dto/response/TodoStatusChangeResponse.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class TodoStatusChangeResponse {
-    private int updateCount;
+    private long updateCount;
 }
